requirements.txt
streamlit>=1.33
langchain>=0.2.0
langchain-google-genai>=0.0.5
langchain-community>=0.2.0
langchain-core>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.6
sqlalchemy>=2.0
